requirements.txt
bert-extractive-summarizer==0.8.1
blis==0.7.4
catalogue==1.0.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cymem==2.0.5
Cython==0.29.23
filelock==3.0.12
Flask==2.0.1
gensim==4.1.2
gunicorn==20.1.0
huggingface-hub==0.0.17
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
murmurhash==1.0.5
nltk==3.6.2
numpy==1.21.2
packaging==21.0
pathy==0.6.0
plac==1.1.3
preshed==3.0.5
pydantic==1.8.2
pyparsing==2.4.7
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.1
sense2vec==1.0.2
six==1.16.0
smart-open==5.2.1
spacy==2.3.7
spacy-legacy==3.0.8
srsly==1.0.5
thinc==7.4.5
threadpoolctl==2.2.0
tokenizers==0.10.3
torch==1.10.0
tqdm==4.62.2
transformers==4.10.3
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.6
wasabi==0.8.2
Werkzeug==2.0.1
